fix(agents): align rehearsal with current build contracts
